refactor MusixmatchLyricsGetter.kt into /external/musixmatch
